Files via API in travelRestController

+++ vs-nextjourney/src/main/java/es/nextjourney/vs_nextjourney/service/ImageService.java
@@ -69,9 +69,9 @@ public class ImageService {
         imageRepository.save(image);
     }
 
-    public Image deleteImageById(long id) {
-        Image image = imageRepository.findById(id).orElseThrow();
-        imageRepository.deleteById(id);
-        return image;
+    public void deleteImageById(long id) {
+        if (imageRepository.existsById(id)) {
+            imageRepository.deleteById(id);
+        }
     }
 }
